25/04/2025

+++ pagina_almacenamiento/delete_file.php
@@ -3,10 +3,16 @@ session_start();
 $id = $_SESSION['id'];
 $data = json_decode(file_get_contents('php://input'), true);
 
-if(isset($data['file'])) {
-    $file = "/mvmup_stor/$id/" . basename($data['file']);
-    
-    if(file_exists($file)) {
+if (isset($data['file'])) {
+    $base_directory = "/mvmup_stor/$id";
+    $file = realpath($base_directory . '/' . $data['file']);
+
+    if (strpos($file, realpath($base_directory)) !== 0) {
+        echo json_encode(['error' => 'Acceso no permitido']);
+        exit;
+    }
+
+    if (file_exists($file)) {
         is_dir($file) ? rmdir($file) : unlink($file);
         echo json_encode(['success' => true]);
     } else {
